major refactor - check notes for details
using component level state for all CRUD actions for courses

context is only used for auth

extra credit challenges are completed:
persist login credentials to localstorage

login in and route to the from in the location if it exsits.

route error mesages from the api for login, registrer user, update course and create course.